pigbabymanager/requirements.txt
django1.10.5
mariadb
MySQL-python.x86_64
webob
